requirements.txt
slackclient>=2.7.1
slackeventsapi>=2.1.0
python-craigslist>0
flask>0